Moved clean_up to clean.c
