+++ include/utils/clean.h
@@ -1,0 +1,1 @@
+void clean_up();
